backend/pyproject.toml
[tool.poetry]
name = "context2task-backend"
version = "1.0.0"
description = "Backend FastAPI para Context2Task - AI-Powered Feature Spec Platform"
authors = ["Context2Task Team"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.116.1"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
pydantic = "^2.5.0"
python-dotenv = "^1.0.0"
httpx = "^0.26.0"
langgraph = "^0.6.0"
langchain-core = "^0.3.0"
typing-extensions = "^4.9.0"
google-generativeai = "^0.8.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
pytest-asyncio = "^0.23.3"
pytest-cov = "^4.1.0"
black = "^24.1.1"
ruff = "^0.1.14"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"


